# server/openapi.yaml
openapi: 3.0.3
info:
  title: NetTechPro API
  description: |
    RESTful API cho website bán thiết bị mạng NetTechPro.
    
    ## Tính năng
    - Quản lý sản phẩm (Products)
    - Quản lý danh mục (Categories)
    - Quản lý thương hiệu (Brands)
    - Quản lý đơn hàng (Orders)
    - Giỏ hàng (Cart)
    - Xác thực (Authentication)
    
    ## Mock Data
    - 8 sản phẩm thiết bị mạng
    - 8 danh mục
    - 10 thương hiệu
    - 3 đơn hàng mẫu
  version: 1.0.0
  contact:
    name: NetTechPro API Support
    email: support@nettechpro.me
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.nettechpro.me
    description: Production server
...
paths:
...
  /api/categories:
    get:
      summary: Lấy danh sách danh mục
      tags: [Categories]
      responses:
        '200':
          description: Danh sách danh mục
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  total:
                    type: integer
    post:
      summary: Tạo danh mục mới
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Danh mục đã được tạo
  /api/categories/{id}:
    get:
      summary: Lấy chi tiết danh mục
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết danh mục
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /api/brands:
    get:
      summary: Lấy danh sách thương hiệu
      tags: [Brands]
      parameters:
        - name: featured
          in: query
          description: Lọc thương hiệu nổi bật
          schema:
            type: boolean
      responses:
        '200':
          description: Danh sách thương hiệu
          content:
            application/json:
              schema:
                type: object
                properties:
                  brands:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'
                  total:
                    type: integer
  /api/brands/{id}:
    get:
      summary: Lấy chi tiết thương hiệu
      tags: [Brands]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết thương hiệu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameVi:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionVi:
          type: string
        productCount:
          type: integer
        order:
          type: integer
    CategoryInput:
      type: object
      required: [name, nameVi, slug]
      properties:
        name:
          type: string
        nameVi:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionVi:
          type: string
    Brand:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logo:
          type: string
        description:
          type: string
        descriptionVi:
          type: string
        country:
          type: string
        productCount:
          type: integer
        featured:
          type: boolean
...